+++ m1sdr.h
@@ -19,7 +19,7 @@ void m1sdr_SetCallback(m1sdr_callback_t *function);
 ao_bool m1sdr_HwPresent(void);
 
 // Backend-specific
-INT16 m1sdr_Init(int sample_rate);
+INT16 m1sdr_Init(char *device, int sample_rate);
 void m1sdr_Exit(void);
 void m1sdr_PlayStart(void);
 void m1sdr_PlayStop(void);