+++ main.c
@@ -148,6 +148,7 @@ int main(int argc, const char *argv[])
 	FILE *file;
 	uint8 *buffer;
 	uint32 size, filesig;
+	char *device = NULL;
 	int noplay = false;
 	int nowave = false;
 
@@ -160,6 +161,9 @@ int main(int argc, const char *argv[])
 	struct argparse_option options[] =
 	{
 		OPT_HELP(),
+		#ifdef WIN32
+		OPT_STRING('d', "device", &device, "name of the playback device"),
+		#endif
 		OPT_BOOLEAN('p', "noplay", &noplay, "don't play back the song"),
 		OPT_BOOLEAN('w', "nowave", &nowave, "don't dump the song to a .wav file"),
 		OPT_END()
@@ -252,7 +256,7 @@ int main(int argc, const char *argv[])
 	signal(SIGINT, intr_handler);
 	if(!noplay)
 	{
-		m1sdr_Init(44100);
+		m1sdr_Init(device, 44100);
 		m1sdr_SetCallback(do_frame);
 		m1sdr_PlayStart();
 		printf("Playing.  ");